% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{analytics_clientstatus_UI}
\alias{analytics_clientstatus_UI}
\title{New or existing client status widget and headings}
\usage{
analytics_clientstatus_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Module generates client status widget and headings.
}
