% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_ui.R
\name{analytics_newcustom_widgets}
\alias{analytics_newcustom_widgets}
\title{Dynamic new customised widgets}
\usage{
analytics_newcustom_widgets(input, output, session)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Module generates widgets for custom variables, defined by the user.
}
