% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{analytics_newcustom_widgets_UI}
\alias{analytics_newcustom_widgets_UI}
\title{Dynamic new customised widgets}
\usage{
analytics_newcustom_widgets_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Module generates widgets for custom variables, defined by the user.
}
