% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{analytics_posttherapy}
\alias{analytics_posttherapy}
\title{New or existing client status widget and headings}
\usage{
analytics_posttherapy(input, output, session, clinician_id,
  selected_client)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Module generates client status widget and headings.
}
