% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{analytics_posttherapy_UI}
\alias{analytics_posttherapy_UI}
\title{Post-Therapy Analytics Widgets}
\usage{
analytics_posttherapy_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Module generates post-therapy analytics widgets
}
