% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{analytics_pretherapy}
\alias{analytics_pretherapy}
\title{New or existing client status widget and headings}
\usage{
analytics_pretherapy(input, output, session)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Module generates client status widget and headings.
}
