% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics 2.R, R/analytics 3.R, R/analytics.R
\name{analytics_pretherapy_UI}
\alias{analytics_pretherapy_UI}
\title{Pre-Therapy Analytics Widgets}
\usage{
analytics_pretherapy_UI(id)

analytics_pretherapy_UI(id)

analytics_pretherapy_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}

\item{id}{String to create a unique namespace.}

\item{id}{String to create a unique namespace.}
}
\description{
Module generates pre-therapy analytics widgets

Module generates pre-therapy analytics widgets

Module generates pre-therapy analytics widgets
}
