% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_ui.R
\name{analytics_widgets}
\alias{analytics_widgets}
\title{Tabpanel widget content for analytics}
\usage{
analytics_widgets(input, output, session, client_status_module)
}
\arguments{
\item{client_status_module}{A module object containing the new/existing client status widget and headings.}

\item{id}{String to create a unique namespace.}
}
\description{
Module generates widgets for entry of basic demographic and clinical information.
}
