% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{analytics_widgets_UI}
\alias{analytics_widgets_UI}
\title{Tabpanel widget content for analytics}
\usage{
analytics_widgets_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Module generates widgets for entry of basic demographic and clinical information.
}
