% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_initial_population.R
\name{apply_initial_population}
\alias{apply_initial_population}
\title{Apply Initial Settings Through Population Selection}
\usage{
apply_initial_population(input, output, session, title, brief_title,
  measure, subscale, population_quantity, populations, sds, means,
  mean_sd_references, reliabilities, reliability_references, cutoff_values,
  cutoff_labels, cutoff_references, cutoff_quantity, items, max_score,
  min_score, description, sample_overview, journal_references,
  existing_data, tabsetpanel_id = "tabset")
}
\arguments{
\item{id}{String to create the namespace}
}
\description{
Select an initial population thereby applying settings for analyses
}
