% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_initial_population.R
\name{apply_initial_population_UI}
\alias{apply_initial_population_UI}
\title{Apply Initial Settings Through Population Selection}
\usage{
apply_initial_population_UI(id)
}
\arguments{
\item{id}{String to create the namespace}
}
\description{
Select an initial population thereby applying settings for analyses
}
