% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate subscale scores.R
\name{calculate_subscale}
\alias{calculate_subscale}
\title{Aggregate subscale scores}
\usage{
calculate_subscale(input, output, session, manual_entry, item_index,
  aggregation_method)
}
\arguments{
\item{item_index}{A list of numeric vectors representing the item indices for each subscale.}

\item{aggregation_method}{A string value indicating the function that should be used to calculate aggregate scores.}

\item{item_scores}{A numeric vector of item scores for the entire scale.}
}
\description{
Calculates the total score of each subscale
}
\examples{
calculate_subscale(item_scores = gad7_scores, item_index = list(c(1,2,3,4,5), c(6,7,8)), aggregation_method = "sum").

}
