% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subscale.R
\name{calculate_subscale_UI}
\alias{calculate_subscale_UI}
\title{Aggregate subscale scores}
\usage{
calculate_subscale_UI(id)
}
\arguments{
\item{id}{String to create the namespace.}
}
\description{
Calculates the total score of each subscale
}
