% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_population.R
\name{change_population}
\alias{change_population}
\title{Update population}
\usage{
change_population(input, output, session, input_population,
  tabsetpanel_id = "tabset")
}
\arguments{
\item{input_population}{String indicating selected population}
}
\description{
Update population widget
}
