% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_population.R
\name{change_population_UI}
\alias{change_population_UI}
\title{Update population}
\usage{
change_population_UI(id)
}
\arguments{
\item{id}{String to create the namespace}
}
\description{
Update population widget
}
