% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_input.R
\name{collect_input}
\alias{collect_input}
\title{Collect subscale input}
\usage{
collect_input(input, output, session, panel_name, manual_entry,
  aggregate_scores, mean_input, sd_input, reliability_input, cutoff_input,
  confidence, method, subscale_number)
}
\arguments{
\item{manual_entry}{A list containing the date and original item scores.}

\item{aggregate_scores}{A list of aggregated subscale scores}

\item{mean_input}{A list containing the mean value and reference (widget inputs)}

\item{sd_input}{A list containing the sd value and reference (widget inputs)}

\item{reliability_input}{A list containing the test-retest value and reference (widget inputs)}

\item{cutoff_input}{A list containing the cutoff values, labels and references (widget inputs)}

\item{confidence}{A numeric value indicating the level of confidence selected for intervals (widget input)}

\item{method}{A string value indicating the reliable change method selected (widget value)}

\item{id}{A string to create a unique namespace.}
}
\description{
Store subscale widget input values in a list
}
