% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_input.R
\name{collect_input_UI}
\alias{collect_input_UI}
\title{Collect subscale input}
\usage{
collect_input_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Store subscale widget input values in a list
}
