% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_all_holding_data.R
\name{combine_all_holding_data}
\alias{combine_all_holding_data}
\title{Combine All Holding Statistics}
\usage{
combine_all_holding_data(input, output, session, holding_statistics_list,
  practice_id)
}
\arguments{
\item{holding_statistics_list}{A list of input lists (one for each subscale).}

\item{practice_id}{String indicating the id of the practice.}
}
\description{
Combine all lists of holding statistics
}
