% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_all_holding_data.R
\name{combine_all_holding_data_UI}
\alias{combine_all_holding_data_UI}
\title{Combine All Holding Statistics}
\usage{
combine_all_holding_data_UI(id)
}
\arguments{
\item{id}{A string to create the namespace.}
}
\description{
Combine all lists of holding statistics
}
