% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_all_input.R
\name{combine_all_input_UI}
\alias{combine_all_input_UI}
\title{Combine All Input}
\usage{
combine_all_input_UI(id)
}
\arguments{
\item{id}{A string to create the namespace.}
}
\description{
Combine all lists of widget input values
}
