% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_ui.R
\name{confidence_level}
\alias{confidence_level}
\title{Confidence Level}
\usage{
confidence_level(input, output, session)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates a confidence interval widget
}
