% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_level.R
\name{confidence_level_UI}
\alias{confidence_level_UI}
\title{Confidence Level}
\usage{
confidence_level_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates a confidence interval widget
}
