% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpaq_20_scale_ui.R
\name{cpaq_20_scale_UI}
\alias{cpaq_20_scale_UI}
\title{CPAQ-20 Scale}
\usage{
cpaq_20_scale_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates the CPAQ-20 for data entry
}
