% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_selected_client_data.R
\name{display_client_data}
\alias{display_client_data}
\title{Select Client}
\usage{
display_client_data(input, output, session, pool, selected_client, measure)
}
\arguments{
\item{pool}{A pooled db object}

\item{selected_client}{A string indicating selected client's unique identifier}

\item{measure}{A string indiating name of the measure}
}
\description{
Select client and display his/her data in abbreviated form.
}
