% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_selected_client_data.R
\name{display_client_data_UI}
\alias{display_client_data_UI}
\title{Select Client}
\usage{
display_client_data_UI(id)
}
\arguments{
\item{id}{A string to creat the namespace}
}
\description{
Select client and display his/her data in abbreviated form.
}
