% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_selected_client_scores.R
\name{display_selected_client_scores}
\alias{display_selected_client_scores}
\title{Display Selected Client Data}
\usage{
display_selected_client_scores(input, output, session,
  selected_client_data)
}
\arguments{
\item{selected_client_data}{A dataframe displaying abbreviated score information for client selected, pulled from db.}
}
\description{
Display abbreviated score information for selected client
}
