% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_selected_client_scores.R
\name{display_selected_client_scores_UI}
\alias{display_selected_client_scores_UI}
\title{Display Selected Client Data}
\usage{
display_selected_client_scores_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Display abbreviated score information for selected client
}
