% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_report.R
\name{download_buttons}
\alias{download_buttons}
\title{Data Submission & Report Generation Widgets}
\usage{
download_buttons(input, output, session)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Creates a submit action button and a report download button to submit data to a database and download a pdf report.
}
