% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_report.R
\name{download_report}
\alias{download_report}
\title{Submit Nested Data to Report}
\usage{
download_report(input, output, session, pool, selected_client, measure,
  manual_entry)
}
\arguments{
\item{pool}{The pooled db connection.}

\item{selected_client}{A string indicating the unique id of the selected client.}

\item{measure}{The name of the psychological measure.}
}
\description{
Wrangle selected_client_data pulled from database and pass it to an R Markdown report
}
