% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_report.R
\name{download_report_UI}
\alias{download_report_UI}
\title{Submit db_database db_data to Report}
\usage{
download_report_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Wrangle db_data pulled from db_database and pass it to an R Markdown report
}
