% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edps_scale.R
\name{edps_scale}
\alias{edps_scale}
\title{EDPS Scale}
\usage{
edps_scale(input, output, session, selected_client)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}
}
\description{
Generates the EDPS for data entry
}
