% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_global_subscale_info.R
\name{export_global_subscale_info}
\alias{export_global_subscale_info}
\title{Export Global Subscale Info}
\usage{
export_global_subscale_info()
}
\description{
Export a global list to Amazon S3 containing the sublists of all subscale parameters.
}
