% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_holding_statistics.R
\name{extract_holding_statistics}
\alias{extract_holding_statistics}
\title{Collect Holding Statistics}
\usage{
extract_holding_statistics(input, output, session, clinician_id, client_id,
  measure, subscale, mean_input, sd_input, reliability_input, confidence,
  method, input_population, cutoff_input, subscale_number)
}
\arguments{
\item{clinician_id}{A string indicating the clinician's unique identifier}

\item{client_id}{A string indicating the client's unique identifier}

\item{measure}{A string indicating the name of the measure.}

\item{subscale}{A string indicating the name of the subscale.}

\item{mean_input}{A list containing the mean value and reference (widget inputs).}

\item{sd_input}{A list containing the sd value and reference (widget inputs).}

\item{reliability_input}{A list containing the test-retest value and reference (widget inputs).}

\item{confidence}{A numeric value indicating the level of confidence selected for intervals (widget input).}

\item{method}{A string value indicating the reliable change method selected (widget value).}

\item{input_population}{A string value indicating the reference population selected by the user.}

\item{cutoff_input}{A list containing the cutoff values, labels and references (widget inputs).}

\item{subscale_number}{A numeric value representing which index (i.e. which subscale) to select from the list of indices.}
}
\description{
Store holding statistics in a list.
}
