% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_holding_statistics.R
\name{extract_holding_statistics_UI}
\alias{extract_holding_statistics_UI}
\title{Collect Holding Statistics}
\usage{
extract_holding_statistics_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Store holding statistics in a list.
}
