% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_severity_ranges.R
\name{find_severity_range}
\alias{find_severity_range}
\title{Find Severity Range}
\usage{
find_severity_range(measure_data)
}
\arguments{
\item{measure_data}{A dataframe indicating the dataframe to send to the db}
}
\description{
Obtain scores and severity ranges for presenting in email
}
