% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_edps_responses_for_email.R
\name{format_edps_responses_for_email_UI}
\alias{format_edps_responses_for_email_UI}
\title{Convert EDPS Measure Responses}
\usage{
format_edps_responses_for_email_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Convert EDPS raw measure responses into a readable table, to be send in an email to the clinician.
}
