% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gad7_item_table.R
\name{format_gad7_responses_for_email}
\alias{format_gad7_responses_for_email}
\title{Convert GAD-7 Measure Responses}
\usage{
format_gad7_responses_for_email(input, output, session, manual_entry)
}
\arguments{
\item{manual_entry}{A list of raw item scores and the date of measure completion.}
}
\description{
Convert GAD-7 raw measure responses into a readable table, to be send in an email to the clinician.
}
