% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gad7_item_table.R
\name{format_gad7_responses_for_email_UI}
\alias{format_gad7_responses_for_email_UI}
\title{Convert GAD-7 Measure Responses}
\usage{
format_gad7_responses_for_email_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Convert GAD-7 raw measure responses into a readable table, to be send in an email to the clinician.
}
