% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gds15_responses_for_email.R
\name{format_gds15_responses_for_email_UI}
\alias{format_gds15_responses_for_email_UI}
\title{Convert GDS-15 Measure Responses}
\usage{
format_gds15_responses_for_email_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Convert GDS-15 raw measure responses into a readable table, to be send in an email to the clinician.
}
