% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pcl5_responses_for_email.R
\name{format_pcl5_responses_for_email}
\alias{format_pcl5_responses_for_email}
\title{Convert PCL-5 Measure Responses}
\usage{
format_pcl5_responses_for_email(input, output, session, pool,
  clinician_email, manual_entry, measure_data, simplified = FALSE)
}
\arguments{
\item{pool}{A pool db connection object}

\item{manual_entry}{A list of raw item scores and the date of measure completion.}

\item{measure_data}{A list of values returned when clinician submits new scale responses. Of list items, only accessing the button value, to trigger the query.}

\item{simplified}{A logical value indicating how the clinician_email string should be parsed (i.e. from being  reactive value or regular string).}
}
\description{
Convert PCL-5 raw measure responses into a readable table, to be send in an email to the clinician.
}
