% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pcl5_responses_for_email.R
\name{format_pcl5_responses_for_email_UI}
\alias{format_pcl5_responses_for_email_UI}
\title{Convert PCL-5 Measure Responses}
\usage{
format_pcl5_responses_for_email_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Convert PCL-5 raw measure responses into a readable table, to be send in an email to the clinician.
}
