% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pdsssr_responses_for_email.R
\name{format_pdsssr_responses_for_email_UI}
\alias{format_pdsssr_responses_for_email_UI}
\title{Convert PDSS-SR Measure Responses}
\usage{
format_pdsssr_responses_for_email_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Convert PDSS-SR raw measure responses into a readable table, to be send in an email to the clinician.
}
