% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gad7_scale.R
\name{gad7_scale}
\alias{gad7_scale}
\title{GAD-7 Scale}
\usage{
gad7_scale(input, output, session, selected_client = NULL,
  simplified = FALSE)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}

\item{simplified}{A boolean value indicating whether scale is being inserted in the remote app.}
}
\description{
Generates the GAD-7 for data entry
}
