% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds15_scale.R
\name{gds15_scale}
\alias{gds15_scale}
\title{GDS-15 Scale}
\usage{
gds15_scale(input, output, session, selected_client = NULL,
  simplified = FALSE)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}

\item{simplified}{A boolean value indicating whether scale is being inserted in the remote app.}
}
\description{
Generates the GDS-15 for data entry
}
