% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds15_scale.R
\name{gds15_scale_UI}
\alias{gds15_scale_UI}
\title{GDS-15 Scale}
\usage{
gds15_scale_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates the GDS-15 for data entry
}
