% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cutoff_widget.R
\name{generate_cutoff_widget_UI}
\alias{generate_cutoff_widget_UI}
\title{Cutoff widget module}
\usage{
generate_cutoff_widget_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates the widget with correct default values for the reliability widget and references.
}
