% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mean_widget.R
\name{generate_mean_widget_UI}
\alias{generate_mean_widget_UI}
\title{Mean widget module}
\usage{
generate_mean_widget_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates the widget with correct default values for the mean widget and references.
}
