% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_derivation.R
\name{generate_reliability_widget}
\alias{generate_reliability_widget}
\title{Reliability widget module}
\usage{
generate_reliability_widget(input, output, session, panel_name,
  subscale_name, population_quantity, populations, input_population, sds,
  means, mean_sd_references, reliabilities, reliability_references,
  cutoffs, cutoff_names, cutoff_references, cutoff_quantity, items,
  max_score, min_score)
}
\arguments{
\item{panel_name}{string (white space allowed) indicating the name of the subscale, to be used as a panel title.}

\item{subscale_name}{A string (underscores should replace white space) indicating the name of the subscale for which the function is being used (e.g. "Anxiety").}

\item{population_quantity}{A numeric value of possible populations from which the user can select.}

\item{populations}{A list of strings (underscores should replace white space) indicating the possible range of populations.}

\item{sds}{A list of numeric values representing the standard deviations for all populations on that subscale.}

\item{means}{A list of numeric values representing the means for all populations on that subscale.}

\item{mean_sd_references}{A list of strings indicating the references for each mean/standard deviation by population.}

\item{reliabilities}{A list of numeric values representing the test-retest reliabilities for all populations on that subscale.}

\item{reliability_references}{A list of strings indicating the references for each reliability value by population.}

\item{cutoffs}{A list of concatenated numeric values representing the cutoff values on this subscale for each population.}

\item{cutoff_names}{A list of concatenated strings indicating the cutoff value descriptors. Use rep() function to multiple by populations.}

\item{cutoff_references}{A list of strings indicating the references for each reliability value by population.}

\item{cutoff_quantity}{A numeric value indicating the number of cutoff scores for the subscale.}
}
\description{
Generates the widget with correct default values for the reliability widget and references.
}
