% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_widget.R
\name{generate_sd_widget_UI}
\alias{generate_sd_widget_UI}
\title{Sd widget module}
\usage{
generate_sd_widget_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates the widget with correct default values for the sd widget and references.
}
