% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simplified_cutoff_widget.R
\name{generate_simplified_cutoff_widget}
\alias{generate_simplified_cutoff_widget}
\title{Simplified Cutoff Widget}
\usage{
generate_simplified_cutoff_widget(input, output, session, subscale_number,
  holding_data)
}
\arguments{
\item{subscale_number}{A number identifiying the subscale (i.e. which row of statistics in the holding table should be selected).}

\item{holding_data}{A dataframe representing the client's existing available data for this measure.}
}
\description{
Generates the simplified application cutoff widget
}
