% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simplified_cutoff_widget.R
\name{generate_simplified_cutoff_widget_UI}
\alias{generate_simplified_cutoff_widget_UI}
\title{Simplified Cutoff Widget}
\usage{
generate_simplified_cutoff_widget_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates the simplified application cutoff widget
}
