% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_mean_widget.R
\name{generate_simplified_mean_widget}
\alias{generate_simplified_mean_widget}
\title{Mean widget module}
\usage{
generate_simplified_mean_widget(input, output, session, holding_data,
  subscale_number, statistic_type)
}
\arguments{
\item{holding_data}{A dataframe representing the client's existing available data for this measure.}

\item{subscale_number}{A number identifiying the subscale (i.e. which row of mean in the holding table should be selected).}

\item{statistic_type}{A string indicating the type of statistic displayed by widget (one of "mean", "sd", "reliability")}
}
\description{
Generates the widget with correct default values for the mean widget and references.
}
