% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_mean_widget.R
\name{generate_simplified_mean_widget_UI}
\alias{generate_simplified_mean_widget_UI}
\title{Simplified Application Mean widget module}
\usage{
generate_simplified_mean_widget_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates the widget with correct default values for the simplified application mean widget and references.
}
