% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_global_subscale_info.R
\name{import_global_subscale_info}
\alias{import_global_subscale_info}
\title{Import Global Subscale Info}
\usage{
import_global_subscale_info()
}
\description{
Import a global list from Amazon S3 containing the sublists of all subscale parameters.
}
