% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_population.R
\name{initiate_settings}
\alias{initiate_settings}
\title{Update population}
\usage{
initiate_settings(input, output, session, input_population,
  tabsetpanel_id = "tabset")
}
\arguments{
\item{input_population}{String indicating selected population}
}
\description{
Update population widget
}
