% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_population.R
\name{initiate_settings_UI}
\alias{initiate_settings_UI}
\title{Update population}
\usage{
initiate_settings_UI(id)
}
\arguments{
\item{id}{String to create the namespace}
}
\description{
Update population widget
}
