% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_method.R
\name{interval_widgets}
\alias{interval_widgets}
\title{Reliable Change Method & Custom Confidence Intervals}
\usage{
interval_widgets(input, output, session, subscale_names)
}
\arguments{
\item{subscale_names}{A vector of character strings representing the names of subscales (underscores should replace white space between words).}
}
\description{
Generates widgets to select reliable change method and custom confidence interval width (if custom intervals are selected)
}
