% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_method_ui.R
\name{interval_widgets_UI}
\alias{interval_widgets_UI}
\title{Reliable Change Method}
\usage{
interval_widgets_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates widget to select reliable change method
}
