% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isi_scale.R
\name{isi_scale}
\alias{isi_scale}
\title{ISI Scale}
\usage{
isi_scale(input, output, session, selected_client)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}
}
\description{
Generates the ISI for data entry
}
