% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_change_variable.R
\name{make_change_variable}
\alias{make_change_variable}
\title{Add Statistically Reliable Change Variable}
\usage{
make_change_variable(subscale_df)
}
\arguments{
\item{subscale_df}{A dataframe containing only data for a particular subscale}
}
\description{
Make new variable (change) that denotes score change and whether such change is statistically reliable
}
