% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ci.R
\name{make_ci}
\alias{make_ci}
\title{Confidence Interval Calculation}
\usage{
make_ci(confidence, se)
}
\arguments{
\item{confidence}{A numeric value representing the user-selected level of confidence for estimation}

\item{se}{A numeric value representing the standard error of prediction for the subscale}
}
\description{
Calculate the confidence interval
}
\examples{
make_ci(1.96, 0.14)

}
