% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_client_dropdown.R
\name{make_client_dropdown}
\alias{make_client_dropdown}
\title{Make Client Dropdown}
\usage{
make_client_dropdown(input, output, session, client_list)
}
\arguments{
\item{client_list}{A reactive list of clients pulled from db}
}
\description{
Render the client dropdown menu
}
