% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_client_dropdown.R
\name{make_client_dropdown_UI}
\alias{make_client_dropdown_UI}
\title{Make Client Dropdown}
\usage{
make_client_dropdown_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Render the client dropdown menu
}
