% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gad7_item_table.R
\name{make_gad7_item_table}
\alias{make_gad7_item_table}
\title{Convert GAD-7 Measure Responses}
\usage{
make_gad7_item_table(input, output, session, manual_entry)
}
\arguments{
\item{manual_entry}{A list of raw item scores and the date of measure completion.}
}
\description{
Convert GAD-7 raw measure responses into a readable table, to be send in an email to the clinician.
}
