% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gad7_item_table.R
\name{make_gad7_item_table_UI}
\alias{make_gad7_item_table_UI}
\title{Convert GAD-7 Measure Responses}
\usage{
make_gad7_item_table_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Convert GAD-7 raw measure responses into a readable table, to be send in an email to the clinician.
}
