% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_selected_subscale_lists.R
\name{make_population_list}
\alias{make_population_list}
\title{Params list generator}
\usage{
make_population_list(list_original_subscale_params, input_population)
}
\arguments{
\item{list_original_subscale_params}{list of original subscale-specific parameters scripts.}
}
\description{
Make a list of population-specific params lists for each type of widget (mean, sd, etc.)
}
