% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pts.R
\name{make_pts}
\alias{make_pts}
\title{Predicted True Score Calculation}
\usage{
make_pts(score, mean, reliability, reliable_change_method)
}
\arguments{
\item{score}{A numeric value representing the aggregate subscale score}

\item{mean}{A numeric value representing the mean score for the subscale}

\item{reliability}{numeric value representing the test-retest reliability for the subscale}

\item{reliable_change_method}{A string representing the method of reliable change calculation}
}
\description{
Calculate the predicted true score
}
\examples{
make_pts(8, 7.2, 0.90, "Nunnally & Bernstein (1994)")

}
