% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_tab .R
\name{make_references_tab_UI}
\alias{make_references_tab_UI}
\title{Make References Tab}
\usage{
make_references_tab_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Module to make references tab
}
