% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_se.R
\name{make_se}
\alias{make_se}
\title{Standard Error Calculation}
\usage{
make_se(sd, reliability, reliable_change_method)
}
\arguments{
\item{sd}{A numeric value representing the standard deviation of the subscale}

\item{reliability}{numeric value representing the test-retest reliability for the subscale}

\item{reliable_change_method}{A string representing the method of reliable change calculation}
}
\description{
Calculate the standard error of prediction
}
\examples{
make_se(.32, .90, "Nunnally & Bernstein (1994)")

}
