% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simplified_sidebar.R
\name{make_simplified_sidebar}
\alias{make_simplified_sidebar}
\title{Make Sidebar}
\usage{
make_simplified_sidebar(input, output, session, start_button_input)
}
\arguments{
\item{start_button_value}{A reactive value containing value of start_button action button.}
}
\description{
Module to make sidebar
}
