% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_sidebar.R
\name{make_simplified_sidebar_UI}
\alias{make_simplified_sidebar_UI}
\title{Make Sidebar}
\usage{
make_simplified_sidebar_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Module to make sidebar
}
