% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_entry.R
\name{manual_data}
\alias{manual_data}
\title{Manual Entry Widgets}
\usage{
manual_data(input, output, session, scale_entry)
}
\arguments{
\item{scale_entry_scores}{The item scores passed from the online scale (if completed).}
}
\description{
Creates layout and widgets for manual data entry
}
