% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_entry.R
\name{manual_data_UI}
\alias{manual_data_UI}
\title{Manual Data Entry Widgets}
\usage{
manual_data_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Creates layout and widgets for manual data entry
}
