% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_method.R
\name{method_widget_UI}
\alias{method_widget_UI}
\title{Reliable Change Method & Custom Confidence Intervals}
\usage{
method_widget_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates widgets to select reliable change method and custom confidence interval width (if custom intervals are selected)
}
