% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_do_call.R
\name{multiply_do_call}
\alias{multiply_do_call}
\title{Multiply do.call()}
\usage{
multiply_do_call(module_id, params_script, input_population)
}
\arguments{
\item{module_id}{List of strings representing ids for each widget}

\item{params_script}{List of original params scripts by subscale}

\item{id}{String to create a unique namespace.}
}
\description{
Multiply the do.call() function to produce widgets without repitition
}
