% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_do_call.R
\name{multiply_do_call_UI}
\alias{multiply_do_call_UI}
\title{Multiply do.call()}
\usage{
multiply_do_call_UI(id, module_id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Multiply the do.call() function to produce widgets without repitition
}
