% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocir_scale.R
\name{ocir_scale}
\alias{ocir_scale}
\title{OCI-R Scale}
\usage{
ocir_scale(input, output, session, selected_client)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}
}
\description{
Generates the OCI-R for data entry
}
