% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_list_maker.R
\name{params_list_maker}
\alias{params_list_maker}
\title{List generator for population paramaters for a subscale}
\usage{
params_list_maker(title, brief_title, measure, subscale,
  population_quantity, populations, input_population, sds, means,
  mean_sd_references, reliabilities, reliability_references, cutoff_values,
  cutoff_labels, cutoff_references, cutoff_quantity, items, max_score,
  min_score, plot_shading_gap, plot_cutoff_label_start,
  plot_cutoff_label_size, description)
}
\arguments{
\item{title}{A string indicating the title of the measure.}

\item{brief_title}{A string indicating the abbreviated title of the measure.}

\item{measure}{A string indicating the measure (no white space).}

\item{subscale}{A string indicating the subscale (no white space).}

\item{population_quantity}{A numeric value of possible populations from which the user can select.}

\item{populations}{A list of strings (underscores should replace white space) indicating the possible range of populations.}

\item{input_population}{A string indicating the population selected by the user.}

\item{sds}{A list of numeric values representing the standard deviations for all populations on that subscale.}

\item{means}{A list of numeric values representing the means for all populations on that subscale.}

\item{mean_sd_references}{A list of strings indicating the references for each mean/standard deviation by population.}

\item{reliabilities}{A list of numeric values representing the test-retest reliabilities for all populations on that subscale.}

\item{reliability_references}{A list of strings indicating the references for each reliability value by population.}

\item{cutoff_references}{A list of strings indicating the references for each reliability value by population.}

\item{cutoff_quantity}{A numeric value indicating the number of cutoff scores for the subscale.}

\item{items}{A numeric vector representing an item index for the subscale.}

\item{max_score}{A numeric value indicating maximum possible score on the subscale.}

\item{min_score}{A numeric value indicating minimum possible score on the subscale.}

\item{plot_shading_gap}{A numeric value subtracted from a cutoff score to indicate the amount of white space between shaded rectangles.}

\item{plot_cutoff_label_start}{A numeric value added to a cutoff score to indicate how far up the cutoff label appears.}

\item{plot_cutoff_label_size}{A numeric value indicating the font size of the cutoff labels.}

\item{description}{A description of subscale's properties, to display in report.}

\item{cutoffs}{A list of concatenated numeric values representing the cutoff values on this subscale for each population.}

\item{cutoff_names}{A list of concatenated strings indicating the cutoff value descriptors. Use rep() function to multiple by populations.}
}
\description{
Outputs the correct list of parameters for the population selected by the user (parameters are then available for default values.
of widgets).
}
