% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdsssr_scale.R
\name{pdsssr_scale_UI}
\alias{pdsssr_scale_UI}
\title{PDSS-SR Scale}
\usage{
pdsssr_scale_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates the PDSS-SR for data entry
}
