% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_cutoff_label.R
\name{position_cutoff_label}
\alias{position_cutoff_label}
\title{Position Cutoff Labels}
\usage{
position_cutoff_label(lower, upper)
}
\arguments{
\item{lower}{Lower border of cutoff shading area}

\item{upper}{Upper border of cutoff shading area}
}
\description{
Position Cutoff Labels on Plots
}
