% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqb_scale.R
\name{pqb_scale}
\alias{pqb_scale}
\title{PQ-B Scale}
\usage{
pqb_scale(input, output, session, selected_client)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}
}
\description{
Generates the PQ-B for data entry
}
