% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_holding_stats.R
\name{read_holding_stats_UI}
\alias{read_holding_stats_UI}
\title{Simplified Application Statistics Read-In}
\usage{
read_holding_stats_UI(id)
}
\arguments{
\item{id}{A string to create the namespace.}
}
\description{
Read in required statistics from holding table in DB, to be used when client completes simplified application measure.
}
