% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_pq_b_missing.R
\name{recode_pq_b_missing}
\alias{recode_pq_b_missing}
\title{Adjust PQ-B missing values vector}
\usage{
recode_pq_b_missing(missing)
}
\arguments{
\item{missing}{A numeric vector indicating the location of missing values}
}
\description{
Implenent PQ-B recording to account for split scorring.
}
