% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_scale_query.R
\name{refresh_scale_query}
\alias{refresh_scale_query}
\title{Trigger Client Data Query On Responses Submission}
\usage{
refresh_scale_query(input, output, session, pool, input_submit_responses,
  measure_data, measure, selected_client)
}
\arguments{
\item{pool}{Pooled db connection}

\item{input_submit_responses}{Reactive object representing id of scale data submission action button}

\item{measure_data}{Reactive object representing the newest ready-for-submission scale data.}

\item{measure}{String representing the scale;}

\item{selected_client}{Reactive object indicating client id that has been selected from client dropdown menu.}
}
\description{
Ensure that client data query is rerun when user submits scale responses, to ensure linear workflow.
}
