% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_scale_query.R
\name{refresh_scale_query_UI}
\alias{refresh_scale_query_UI}
\title{Trigger Client Data Query On Responses Submission}
\usage{
refresh_scale_query_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Ensure that client data query is rerun when user submits scale responses, to ensure linear workflow.
}
