% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_calcs_ui.R
\name{reliability_calc}
\alias{reliability_calc}
\title{Test-Retest Reliability Derivation}
\usage{
reliability_calc(input, output, session)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Calculates test-retest reliability from T- or F-statistic
}
