% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_calc.R
\name{reliability_calc_UI}
\alias{reliability_calc_UI}
\title{Test-Retest Reliability Derivation}
\usage{
reliability_calc_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Calculates test-retest reliability from T- or F-statistic
}
