% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_client_dropdown.R
\name{render_client_dropdown}
\alias{render_client_dropdown}
\title{Render Client Dropdown}
\usage{
render_client_dropdown(input, output, session, pool, clinician_id)
}
\arguments{
\item{pool}{A pool db connection object}

\item{clinician_id}{Clinician's unique identifier}
}
\description{
#Render client dropdown menu
}
