% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_client_dropdown.R
\name{render_client_dropdown_UI}
\alias{render_client_dropdown_UI}
\title{Render Client Dropdown}
\usage{
render_client_dropdown_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
#Render client dropdown menu
}
