% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_availability.R
\name{report_availability_message}
\alias{report_availability_message}
\title{Report Availability Message}
\usage{
report_availability_message(input, output, session, selected_client_data)
}
\arguments{
\item{selected_client_data}{Dataframe of client's outcomes}
}
\description{
Make message about data availability for report
}
