% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_availability.R
\name{report_availability_message_UI}
\alias{report_availability_message_UI}
\title{Report Availability Message}
\usage{
report_availability_message_UI(id)
}
\arguments{
\item{id}{String to create the namespace}
}
\description{
Make message about data availability for report
}
