% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_client_name.R
\name{retrieve_client_name}
\alias{retrieve_client_name}
\title{Retrieve Client Name}
\usage{
retrieve_client_name(input, output, session, pool,
  input_retrieve_client_data, selected_client)
}
\arguments{
\item{pool}{A pooled db connection}

\item{input_retrieve_client_data}{A reactive value: the value of client selection button}

\item{selected_client}{A reactive value containing the id of the selected client}
}
\description{
Retrive name to display thoughout pages of parent app
}
