% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_client_name.R
\name{retrieve_client_name_UI}
\alias{retrieve_client_name_UI}
\title{Retrieve Client Name}
\usage{
retrieve_client_name_UI(id)
}
\arguments{
\item{id}{A string to creat the namespace}
}
\description{
Retrive name to display thoughout pages of parent app
}
