% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_display_selected_client.R
\name{retrieve_selected_client}
\alias{retrieve_selected_client}
\title{Retrieve & Show Client Data}
\usage{
retrieve_selected_client(input, output, session, pool, selected_client,
  measure)
}
\arguments{
\item{selected_client}{Id of the client selected in dropdown}

\item{measure}{Measure for which scores are pulled in from db}
}
\description{
Query database to retrive client scores and display in table
}
