% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_display_selected_client.R
\name{retrieve_selected_client_UI}
\alias{retrieve_selected_client_UI}
\title{Retrieve Selected Client Data}
\usage{
retrieve_selected_client_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Query database to retrive abbreviated client score information
}
