% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_scale.R
\name{sds_scale}
\alias{sds_scale}
\title{SDS Scale}
\usage{
sds_scale(input, output, session, selected_client)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}
}
\description{
Generates the SDS for data entry
}
