% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_dropdown_select.R
\name{select_client}
\alias{select_client}
\title{Client Dropdown & Select}
\usage{
select_client(input, output, session, pool, clinician_id)
}
\arguments{
\item{pool}{Pool database connection}

\item{clinician_id}{Id of the clinician passed from Shiny Proxy login}
}
\description{
Create client dropdown menu and enable client selection from db
}
