% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_dropdown_select.R
\name{select_client_UI}
\alias{select_client_UI}
\title{Client Dropdown & Select}
\usage{
select_client_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Create client dropdown menu and enable client selection from db
}
