% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_population.R
\name{select_population}
\alias{select_population}
\title{Update Population}
\usage{
select_population(input, output, session)
}
\arguments{
\item{selected_client_data}{Data for the selected patient pulled from the db. The required value is selected_client_data()$population}
}
\description{
Update population selection widget with value chosen at registration
}
