% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_population.R
\name{select_population_UI}
\alias{select_population_UI}
\title{Population Selection}
\usage{
select_population_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Selected a population matching client characteristics
}
