% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_client_name.R
\name{show_client_name_UI}
\alias{show_client_name_UI}
\title{Make Client Name For Display}
\usage{
show_client_name_UI(id)
}
\arguments{
\item{id}{A string to create the namespace.}
}
\description{
Pull name from db to display on app tabpanels
}
