% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_population_message.R
\name{show_population_message}
\alias{show_population_message}
\title{Selected Population Message}
\usage{
show_population_message(input, output, session, input_population)
}
\arguments{
\item{input_population}{String indicating selected population}
}
\description{
Show a message indicating the population that was selected
}
