% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_population_message.R
\name{show_population_message_UI}
\alias{show_population_message_UI}
\title{Selected Population Message}
\usage{
show_population_message_UI(id)
}
\arguments{
\item{id}{String to create the namespace}
}
\description{
Show a message indicating the population that was selected
}
