% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sias_scale.R
\name{sias_scale}
\alias{sias_scale}
\title{SIAS Scale}
\usage{
sias_scale(input, output, session, selected_client)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}
}
\description{
Generates the SIAS for data entry
}
