% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_collect_holding_input.R
\name{simplified_collect_holding_input_UI}
\alias{simplified_collect_holding_input_UI}
\title{Collect Simplified Application Subscale Input}
\usage{
simplified_collect_holding_input_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Collect simplified application subscale data for writing to db.
}
