% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_confidence_level.R
\name{simplified_confidence_level}
\alias{simplified_confidence_level}
\title{Confidence Level}
\usage{
simplified_confidence_level(input, output, session, holding_data)
}
\arguments{
\item{holding_data}{A dataframe representing the client's statistics in the DB holding table.}
}
\description{
Generates a confidence interval widget
}
