% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifed application confidence.R
\name{simplified_confidence_level_UI}
\alias{simplified_confidence_level_UI}
\title{Simplified Application Confidence Level}
\usage{
simplified_confidence_level_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates a confidence interval widget
}
