% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_gad_7_scale.R
\name{simplified_gad7_scale}
\alias{simplified_gad7_scale}
\title{GAD-7 Scale}
\usage{
simplified_gad7_scale(input, output, session, holding_data)
}
\arguments{
\item{holding_data}{A dataframe containing statistics pulled from the holding table in DB.}
}
\description{
Generates the GAD-7 for data entry
}
