% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_gad_7_scale.R
\name{simplified_gad7_scale_UI}
\alias{simplified_gad7_scale_UI}
\title{Simplified GAD-7 Scale}
\usage{
simplified_gad7_scale_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Dynamically enerates the GAD-7 for simplified application.
}
