% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_manual_entry.R
\name{simplified_manual_data}
\alias{simplified_manual_data}
\title{Simplified Manual Data Entry Widgets}
\usage{
simplified_manual_data(input, output, session, scale_entry)
}
\arguments{
\item{scale_entry}{The item scores passed from the online scale (if completed).}
}
\description{
Creates layout and widgets for simple application manual data entry
}
