% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_manual_data.R
\name{simplified_manual_data_UI}
\alias{simplified_manual_data_UI}
\title{Simplified Manual Data Entry Widgets}
\usage{
simplified_manual_data_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Creates layout and widgets for simple application manual data entry
}
