% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_reliable_change_widget.R
\name{simplified_method_widget}
\alias{simplified_method_widget}
\title{Reliable Change Method & Custom Confidence Intervals}
\usage{
simplified_method_widget(input, output, session, holding_data)
}
\arguments{
\item{holding_data}{A dataframe representing the client's statistics in the DB holding table.}
}
\description{
Generates widgets to select reliable change method and custom confidence interval width (if custom intervals are selected)
}
