% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_reliable_change_widget.R
\name{simplified_method_widget_UI}
\alias{simplified_method_widget_UI}
\title{Simplifed Application Reliable Change Method}
\usage{
simplified_method_widget_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates simplified application widgets to select reliable change method.
}
