% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_show_name.R
\name{simplified_show_name}
\alias{simplified_show_name}
\title{Make Client Name For Display}
\usage{
simplified_show_name(input, output, session, holding_data)
}
\arguments{
\item{client_name_for_display}{A reactive object (string) containing the client's first and last names.}

\item{input_retrive_client_data}{A reactive value: the value of client selection button.}
}
\description{
Pull name from db to display on app tabpanels
}
