% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_show_name.R
\name{simplified_show_name_UI}
\alias{simplified_show_name_UI}
\title{Make Client Name For Display}
\usage{
simplified_show_name_UI(id)
}
\arguments{
\item{id}{A string to create the namespace.}
}
\description{
Pull name from db to display on app tabpanels
}
