% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps_scale.R
\name{sps_scale}
\alias{sps_scale}
\title{SPS Scale}
\usage{
sps_scale(input, output, session, selected_client)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}
}
\description{
Generates the SPS for data entry
}
