% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssd12_scale.R
\name{ssd12_scale}
\alias{ssd12_scale}
\title{SSD-12 Scale}
\usage{
ssd12_scale(input, output, session, selected_client = NULL,
  simplified = FALSE)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}

\item{simplified}{A boolean value indicating whether scale is being inserted in the remote app.}
}
\description{
Generates the SSD-12 for data entry
}
