% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssd12_scale.R
\name{ssd12_scale_UI}
\alias{ssd12_scale_UI}
\title{SSD-12 Scale}
\usage{
ssd12_scale_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Generates the SSD-12 for data entry
}
