% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_ui.R
\name{stats_tabwidgets}
\alias{stats_tabwidgets}
\title{Tabpanel widget content for mean, sd, test-retest reliability and cutoff values}
\usage{
stats_tabwidgets(input, output, session, subscale_names, stats_df, refs_df,
  pop)
}
\arguments{
\item{subscale_names}{A list of strings representing the names the subscales}

\item{stats_df}{A dataframe. Column names indicate possible populations (e.g. veteran). Integers within
each column represent subscale values. Subscale values in columns must match the order in which subscales
are listed in the subscale_names argument.}

\item{refs_df}{A dataframe. Column names indicate possible populations (e.g. veteran). Strings within each
column represent references corresponding to each subscale value. Subscale references in columns must match the order in which subscales
are listed in the subscale_names argument.}

\item{pop}{A reactive value (string) derived from a widget.}
}
\description{
Module generates the widgets (and associated references) required for inputting mean,
sd, reliability and cutoff values.
}
