% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_ui.R
\name{stats_tabwidgets_UI}
\alias{stats_tabwidgets_UI}
\title{Tabpanel widget content for mean, sd, test-retest reliability and cutoff values}
\usage{
stats_tabwidgets_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Module generates the widgets (and associated references) required for inputting mean,
sd, reliability and cutoff values.
}
