% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_pretherapy_analytics.R
\name{submit_pretherapy_analytics}
\alias{submit_pretherapy_analytics}
\title{Submit Pretherapy Analytics}
\usage{
submit_pretherapy_analytics(input, output, session, pool, clinician_id)
}
\description{
Nested module to create pretherapy analytics dataframe and send it to db
}
