% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_pretherapy_analytics.R
\name{submit_pretherapy_analytics_UI}
\alias{submit_pretherapy_analytics_UI}
\title{Submit Pretherapy Analytics}
\usage{
submit_pretherapy_analytics_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Nested module to create pretherapy analytics dataframe and send it to db
}
