% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{t_rel}
\alias{t_rel}
\title{T Rel}
\usage{
t_rel(calc_sd, calc_retest_sd, calc_retest_m, calc_m, calc_t, calc_n,
  digits = 2)
}
\arguments{
\item{calc_sd}{numeric representing the standard deviation}

\item{calc_n}{}
}
\description{
Calculates the something something
}
