% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn_too_many_responses.R
\name{warn_too_many_responses}
\alias{warn_too_many_responses}
\title{Indicate Excessive Item Response}
\usage{
warn_too_many_responses(session, item_too_long)
}
\arguments{
\item{session}{A session object.}

\item{item_too_long}{A logical vector indicating whether an input is greater than length one.}
}
\description{
Make sweet alert indicating endorsement of too many items
}
