% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_items_for_db.R
\name{wrangle_items_for_db}
\alias{wrangle_items_for_db}
\title{Make Item-Level Entry For Item Table In Database}
\usage{
wrangle_items_for_db(input, output, session, manual_entry, measure_data)
}
\arguments{
\item{manual_entry}{A list of values returned when clinician submits new scale responses. Of list items, only accessing the button value, to trigger the query.}

\item{measure_data}{A dataframe indicating the dataframe to send to the db}
}
\description{
Wrangle the individual item scores and other key information for writing to the item table in the database.
}
