% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_items_for_db.R
\name{wrangle_items_for_db_UI}
\alias{wrangle_items_for_db_UI}
\title{Make Item-Level Entry For Item Table In Database}
\usage{
wrangle_items_for_db_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Wrangle the individual item scores and other key information for writing to the item table in the database.
}
