% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_measure_data_to_db 2.R,
%   R/write_measure_data_to_db.R
\name{write_measure_data_to_db}
\alias{write_measure_data_to_db}
\title{Write client's measure data to db}
\usage{
write_measure_data_to_db(input, ouput, session, pool, measure_data,
  manual_entry)

write_measure_data_to_db(input, ouput, session, pool, measure_data,
  manual_entry)
}
\arguments{
\item{pool}{A pooled db connection}

\item{measure_data}{A dataframe indicating the dataframe to send to the db}

\item{manual_entry}{A list of values returned when clinician submits new scale responses. Of list items, only accessing the button value, to trigger the query.}

\item{measure_data}{A dataframe indicating the dataframe to send to the db}

\item{manual_entry}{A list of values returned when clinician submits new scale responses. Of list items, only accessing the button value, to trigger the query.}
}
\description{
Write the client's new scores and associated statistics to the db

Write the client's new scores and associated statistics to the db
}
