% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_measure_data_to_db.R
\name{write_measure_data_to_db_UI}
\alias{write_measure_data_to_db_UI}
\title{Write client's measure data to db}
\usage{
write_measure_data_to_db_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Write the client's new scores and associated statistics to the db
}
