% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_posttherapy_to_db.R
\name{write_posttherapy_to_db}
\alias{write_posttherapy_to_db}
\title{Write post-therapy analytics}
\usage{
write_posttherapy_to_db(input, output, session, pool,
  analytics_posttherapy)
}
\arguments{
\item{pool}{A pooled db connection.}

\item{analytics_posttherapy}{A dataframe containing post-therapy analytics data.}
}
\description{
Write post-therapy analytics data to the db
}
