% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_posttherapy_to_db.R
\name{write_posttherapy_to_db_UI}
\alias{write_posttherapy_to_db_UI}
\title{Write post-therapy analytics}
\usage{
write_posttherapy_to_db_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Write post-therapy analytics data to the db
}
