% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pretherapy_to_db.R
\name{write_pretherapy_to_db_UI}
\alias{write_pretherapy_to_db_UI}
\title{Write pre-therapy analytics}
\usage{
write_pretherapy_to_db_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Write pre-therapy analytics data to the db
}
