% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_statistics_to_holding.R
\name{write_statistics_to_holding}
\alias{write_statistics_to_holding}
\title{Write client's statistics to holding table}
\usage{
write_statistics_to_holding(input, ouput, session, pool, holding_data)
}
\arguments{
\item{pool}{A pooled database connection.}

\item{holding_data}{A dataframe indicating the dataframe to send to the db.}
}
\description{
Write the client's statistics (selected by clinician) to holding table in DB if sending measure via email link.
}
