% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_statistics_to_holding.R
\name{write_statistics_to_holding_UI}
\alias{write_statistics_to_holding_UI}
\title{Write client's statistics to holding table}
\usage{
write_statistics_to_holding_UI(id)
}
\arguments{
\item{id}{A string to create the namespace}
}
\description{
Write the client's statistics (selected by clinician) to holding table in DB if sending measure via email link.
}
