% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_pretherapy_posttherapy.R
\name{analytics_posttherapy}
\alias{analytics_posttherapy}
\title{New or existing client status widget and headings}
\usage{
analytics_posttherapy(input, output, session, clinician_id,
  selected_client, assessment_stage)
}
\arguments{
\item{clinician_id}{String indicating clinician's unique identifier.}

\item{selected_client}{String indicating client's unique identifier.}

\item{id}{String to create a unique namespace.}
}
\description{
Module generates client status widget and headings.
}
