% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_pretherapy_posttherapy.R
\name{analytics_pretherapy}
\alias{analytics_pretherapy}
\title{New or existing client status widget and headings}
\usage{
analytics_pretherapy(input, output, session, clinician_id)
}
\arguments{
\item{clinician_id}{String indicating clinician's unique identifier.}

\item{id}{String to create a unique namespace.}
}
\description{
Module generates client status widget and headings.
}
