% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics 2.R, R/analytics 3.R, R/analytics.R
\name{analytics_pretherapy}
\alias{analytics_pretherapy}
\title{New or existing client status widget and headings}
\usage{
analytics_pretherapy(input, output, session, clinician_id)

analytics_pretherapy(input, output, session, clinician_id)

analytics_pretherapy(input, output, session, clinician_id)
}
\arguments{
\item{id}{String to create a unique namespace.}

\item{id}{String to create a unique namespace.}

\item{id}{String to create a unique namespace.}
}
\description{
Module generates client status widget and headings.

Module generates client status widget and headings.

Module generates client status widget and headings.
}
