% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_initial_population.R
\name{apply_initial_population}
\alias{apply_initial_population}
\title{Apply Initial Settings Through Population Selection}
\usage{
apply_initial_population(input, output, session, title, brief_title,
  measure, subscale, population_quantity, populations, sds, means,
  mean_sd_references, reliabilities, reliability_references, cutoff_values,
  cutoff_labels, cutoff_references, cutoff_quantity, items, max_score,
  min_score, description, sample_overview, journal_references,
  existing_data, tabsetpanel_id = "tabset")
}
\arguments{
\item{title}{A string (white space allowed) indicating the name of the subscale, to be used as a panel title.}

\item{brief_title}{An abbreviated title or acronym.}

\item{measure}{A string indicating the global measure.}

\item{subscale}{A string (underscores should replace white space) indicating the name of the subscale for which the function is being used (e.g. "Anxiety").}

\item{population_quantity}{A numeric value of possible populations from which the user can select.}

\item{populations}{A list of strings (underscores should replace white space) indicating the possible range of populations.}

\item{sds}{A list of numeric values representing the standard deviations for all populations on that subscale.}

\item{means}{A list of numeric values representing the means for all populations on that subscale.}

\item{mean_sd_references}{A list of strings indicating the references for each mean/standard deviation by population.}

\item{reliabilities}{A list of numeric values representing the test-retest reliabilities for all populations on that subscale.}

\item{reliability_references}{A list of strings indicating the references for each reliability value by population.}

\item{cutoff_values}{A list of concatenated numeric values representing the cutoff values on this subscale for each population.}

\item{cutoff_labels}{A list of concatenated strings indicating the cutoff value descriptors. Use rep() function to multiple by populations.}

\item{cutoff_references}{A list of strings indicating the references for each reliability value by population.}

\item{cutoff_quantity}{A numeric value indicating the number of cutoff scores for the subscale.}

\item{items}{A numeric vector representing an item index for the subscale.}

\item{max_score}{A numeric value indicating maximum possible score on the subscale.}

\item{min_score}{A numeric value indicating minimum possible score on the subscale.}

\item{description}{A string indicating a description of subscale's properties, to display in report}

\item{sample_overview}{A string indicating a description of each research sample (different for each population)}

\item{journal_references}{The string indicating a full reference for each research sample for mean and sd.}

\item{tabsetpanel_id}{A string indicating a panel id to allow automatic switching between panels.}

\item{input_population}{A string (reactive value) representing the name of the selected population.}
}
\description{
Select an initial population thereby applying settings for analyses
}
