% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_input.R
\name{collect_input}
\alias{collect_input}
\title{Collect subscale input}
\usage{
collect_input(input, output, session, clinician_id, client_id, measure,
  subscale, manual_entry, aggregate_scores, mean_input, sd_input,
  reliability_input, confidence, method, input_population, cutoff_input,
  subscale_number)
}
\arguments{
\item{client_id}{A string indicating the client's unique id}

\item{manual_entry}{A list containing the date and original item scores.}

\item{aggregate_scores}{A list of aggregated subscale scores}

\item{mean_input}{A list containing the mean value and reference (widget inputs)}

\item{sd_input}{A list containing the sd value and reference (widget inputs)}

\item{reliability_input}{A list containing the test-retest value and reference (widget inputs)}

\item{confidence}{A numeric value indicating the level of confidence selected for intervals (widget input)}

\item{method}{A string value indicating the reliable change method selected (widget value)}

\item{input_population}{A string value indicating the reference population selected by the user}

\item{cutoff_input}{A list containing the cutoff values, labels and references (widget inputs)}
}
\description{
Store subscale widget input values in a list
}
