% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_level.R
\name{confidence_level}
\alias{confidence_level}
\title{Confidence Level}
\usage{
confidence_level(input, output, session, existing_data)
}
\arguments{
\item{existing_data}{A dataframe representing the client's existing available data for this measure.}
}
\description{
Generates a confidence interval widget
}
