% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_report.R
\name{download}
\alias{download}
\title{Data Submission & Report Generation Widgets}
\usage{
download(input, output, session, all_input)
}
\arguments{
\item{current_data}{Item scores and date}

\item{mean_output_list_1}{Mean widgets output}
}
\description{
Creates a submit action button and a report download button to submit data to a database and download a pdf report.
}
