% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_report.R
\name{download_report}
\alias{download_report}
\title{Submit Nested Data to Report}
\usage{
download_report(input, output, session, pool, selected_client,
  global_subscale_info, most_recent_client_data)
}
\arguments{
\item{pool}{The pooled db connection.}

\item{selected_client}{A string indicating the unique id of the selected client.}

\item{global_subscale_info}{A list of lists containing required statistics for each outcome measure subscale.}

\item{most_recent_client_data}{A dataframe representing the most recently queried data (across all measures) for this patient.}
}
\description{
Wrangle selected_client_data pulled from database and pass it to an R Markdown report
}
