% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_report.R
\name{download_report}
\alias{download_report}
\title{Submit db_database db_data to Report}
\usage{
download_report(input, output, session, db_data)
}
\arguments{
\item{db_data}{An unnested dataframe pulled from the database containing all timepoint scores for an individual on a specific measure.}
}
\description{
Wrangle db_data pulled from db_database and pass it to an R Markdown report
}
