% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_report.R
\name{download_report}
\alias{download_report}
\title{Submit Nested Data to Report}
\usage{
download_report(input, output, session, selected_client_data,
  selected_client, pool)
}
\arguments{
\item{selected_client_data}{An unnested dataframe pulled from the database containing all timepoint scores for an individual on a specific measure.}
}
\description{
Wrangle selected_client_data pulled from database and pass it to an R Markdown report
}
