% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_report.R
\name{download_report_UI}
\alias{download_report_UI}
\title{Submit Nested Data to Report}
\usage{
download_report_UI(id)
}
\arguments{
\item{id}{String to create a unique namespace.}
}
\description{
Wrangle selected_client_data pulled from database and pass it to an R Markdown report
}
