% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_entry.R
\name{manual_data}
\alias{manual_data}
\title{Manual Entry Widgets}
\usage{
manual_data(input, output, session, scale_entry, do_recode = FALSE,
  items_to_recode = NULL, recoding_key = NULL)
}
\arguments{
\item{scale_entry}{The item scores passed from the online scale (if completed).}

\item{do_recode}{A logical, indicating whether or not to apply the recoding function.}

\item{items_to_recode}{A numeric vector indicating the items that should be recoded, of the form c(3,5,6) etc.}

\item{recoding_key}{A string indicating the recoding pattern to use, of the form '0=;1=5;2=4;3=3;4=2;5=1;6=0' etc.}
}
\description{
Creates layout and widgets for manual data entry
}
