% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_method.R
\name{method_widget}
\alias{method_widget}
\title{Reliable Change Method & Custom Confidence Intervals}
\usage{
method_widget(input, output, session, existing_data)
}
\arguments{
\item{existing_data}{A dataframe indicating the client's existing available data for this measure.}
}
\description{
Generates widgets to select reliable change method and custom confidence interval width (if custom intervals are selected)
}
