% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phq9_scale.R
\name{phq9_scale}
\alias{phq9_scale}
\title{PHQ-9 Scale}
\usage{
phq9_scale(input, output, session, selected_client = NULL,
  simplified = FALSE)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}

\item{simplified}{A boolean value indicating whether scale is being inserted in the remote app.}
}
\description{
Generates the PHQ-9 for data entry
}
