% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subscale.R
\name{plot_subscale}
\alias{plot_subscale}
\title{Plot Subscale}
\usage{
plot_subscale(subscale_df, subscale_info)
}
\arguments{
\item{subscale_df}{A dataframe containing only data for a particular subscale}

\item{subscale_info}{A list of subscale parameters}
}
\description{
Plot an individual subscale
}
\examples{
plot_subscale(subscale_df = subscale_df, subscale_info= psychlytx::gad7_params)
}
