% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_holding_stats.R
\name{read_holding_stats}
\alias{read_holding_stats}
\title{Simplified Application Statistics Read-In}
\usage{
read_holding_stats(input, output, session, pool, measure,
  tabsetpanel_id = "tabset", client_id)
}
\arguments{
\item{pool}{A pooled db object.}

\item{measure}{A string indiating name of the measure.}

\item{tabsetpanel_id}{A string to allow automated switch from sign-in tabpanel to questionnaire tabpanel.}
}
\description{
Read in required statistics from holding table in DB, to be used when client completes simplified application measure.
}
