% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_holding_stats.R
\name{read_holding_stats}
\alias{read_holding_stats}
\title{Select Client}
\usage{
read_holding_stats(input, output, session, pool, measure)
}
\arguments{
\item{pool}{A pooled db object.}

\item{measure}{A string indiating name of the measure.}
}
\description{
Select client and display his/her data in abbreviated form.
}
