% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_reliability.R
\name{reliability_calc}
\alias{reliability_calc}
\title{Test-Retest Reliability Derivation}
\usage{
reliability_calc(input, output, session)
}
\description{
Calculates test-retest reliability from T- or F-statistic
}
