% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplfied_gad7_scale.R
\name{simplified_gad7_scale}
\alias{simplified_gad7_scale}
\title{GAD-7 Scale}
\usage{
simplified_gad7_scale(input, output, session, client_id)
}
\arguments{
\item{selected_client}{A string indicating the unique id of the selected client.}
}
\description{
Generates the GAD-7 for data entry
}
